Fix unit tests and transfer back to GridTests.cs
